Implement Lottie-animated splash screen and refactor project structure
This commit introduces a custom splash screen using the AndroidX Core Splashscreen library and Lottie animations, while reorganizing the package structure to move meal-related data and domain logic into a dedicated `meals` module.

- **Splash Screen**:
    - Added `SplashActivity` with Lottie animation support (`splash_anim.json`).
    - Configured `Theme.App.Splash` using `androidx.core:core-splashscreen`.
    - Implemented full-screen layout and smooth transition to `MainActivity`.
    - Added relevant string resources and color definitions for the splash UI.
- **Refactoring**:
    - Relocated `data`, `domain`, and `presentation` layers into a new `meals` sub-package (e.g., `com.example.recipe_android_project.meals.data`).
    - Updated package declarations and imports across all moved classes (Mappers, Entities, DTOs, and Models).
- **Resources**:
    - Added `loading_anim.json` and `splash_anim.json` to raw resources.
    - Defined new theme `Theme.RecipeAndroidProject` and updated `AndroidManifest.xml` to set `SplashActivity` as the entry point.
- **Dependencies**:
    - Added `androidx.core:core-splashscreen:1.2.0` to `libs.versions.toml` and `app/build.gradle.kts`.
