+++ app/build.gradle.kts
@@ -46,7 +46,7 @@ dependencies {
     implementation(libs.androidx.room.runtime)
     annotationProcessor(libs.androidx.room.compiler)
     implementation(libs.lottie.dep)
-
+    implementation(libs.androidx.core.splashscreen)
     implementation(libs.constraintlayout)
     testImplementation(libs.junit)
     androidTestImplementation(libs.ext.junit)
