--- app/src/main/java/com/example/recipe_android_project/data/repository/repo.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.data.repository;
-
-public class repo {
-}
