--- app/src/main/java/com/example/recipe_android_project/domain/model/User.java
@@ -1,175 +1,0 @@
-package com.example.recipe_android_project.domain.model;
-
-import java.io.Serializable;
-
-public class User implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    private String id;
-    private String fullName;
-    private String email;
-    private String password;
-    private boolean isLoggedIn;
-    private long createdAt;
-    private long updatedAt;
-
-    public User() {
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-    public User(String fullName, String email, String password) {
-        this.fullName = fullName;
-        this.email = email;
-        this.password = password;
-        this.createdAt = System.currentTimeMillis();
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-    // ==================== HELPER METHODS ====================
-
-    public boolean isValid() {
-        return isValidFullName() && isValidEmail() && isValidPassword();
-    }
-
-    public boolean isValidFullName() {
-        return fullName != null && fullName.trim().length() >= 2;
-    }
-
-    public boolean isValidEmail() {
-        if (email == null || email.isEmpty()) {
-            return false;
-        }
-        String emailRegex = "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$";
-        return email.matches(emailRegex);
-    }
-
-    public boolean isValidPassword() {
-        return password != null && password.length() >= 6;
-    }
-
-    public String getFirstName() {
-        if (fullName == null || fullName.isEmpty()) {
-            return "";
-        }
-        String[] parts = fullName.trim().split("\\s+");
-        return parts[0];
-    }
-
-    public String getLastName() {
-        if (fullName == null || fullName.isEmpty()) {
-            return "";
-        }
-        String[] parts = fullName.trim().split("\\s+");
-        if (parts.length > 1) {
-            return parts[parts.length - 1];
-        }
-        return "";
-    }
-
-    public String getInitials() {
-        if (fullName == null || fullName.isEmpty()) {
-            return "";
-        }
-        String[] parts = fullName.trim().split("\\s+");
-        if (parts.length >= 2) {
-            return ("" + parts[0].charAt(0) + parts[parts.length - 1].charAt(0)).toUpperCase();
-        }
-        return ("" + parts[0].charAt(0)).toUpperCase();
-    }
-
-
-
-    // ==================== GETTERS & SETTERS ====================
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-
-    public boolean isLoggedIn() {
-        return isLoggedIn;
-    }
-
-    public void setLoggedIn(boolean loggedIn) {
-        isLoggedIn = loggedIn;
-        this.updatedAt = System.currentTimeMillis();
-    }
-
-    public long getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(long createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public long getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(long updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        User user = (User) o;
-        if (id != null) {
-            return id.equals(user.id);
-        }
-        return email != null && email.equals(user.email);
-    }
-
-    @Override
-    public int hashCode() {
-        if (id != null) {
-            return id.hashCode();
-        }
-        return email != null ? email.hashCode() : 0;
-    }
-
-    @Override
-    public String toString() {
-        return "User{" +
-                "id='" + id + '\'' +
-                ", fullName='" + fullName + '\'' +
-                ", email='" + email + '\'' +
-                ", isLoggedIn=" + isLoggedIn +
-                '}';
-    }
-}
