--- app/src/main/java/com/example/recipe_android_project/domain/usecases/usecase.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.domain.usecases;
-
-public class usecase {
-}
