+++ app/src/main/java/com/example/recipe_android_project/meals/data/local/local.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.data.local;
+
+public class local {
+}
