+++ app/src/main/java/com/example/recipe_android_project/meals/data/mapper/mapper.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.data.mapper;
+
+public class mapper {
+}
