+++ app/src/main/java/com/example/recipe_android_project/meals/data/remote/api/api.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.data.remote.api;
+
+public class api {
+}
