+++ app/src/main/java/com/example/recipe_android_project/meals/data/repository/repo.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.data.repository;
+
+public class repo {
+}
