+++ app/src/main/java/com/example/recipe_android_project/meals/domain/usecases/usecase.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.domain.usecases;
+
+public class usecase {
+}
