+++ app/src/main/java/com/example/recipe_android_project/meals/presentation/base/base.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.presentation.base;
+
+public class base {
+}
