+++ app/src/main/java/com/example/recipe_android_project/meals/presentation/widgets/widgets.java
@@ -1,0 +1,4 @@
+package com.example.recipe_android_project.meals.presentation.widgets;
+
+public class widgets {
+}
