--- app/src/main/java/com/example/recipe_android_project/presentation/base/base.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.presentation.base;
-
-public class base {
-}
