--- app/src/main/java/com/example/recipe_android_project/presentation/widgets/widgets.java
@@ -1,4 +1,0 @@
-package com.example.recipe_android_project.presentation.widgets;
-
-public class widgets {
-}
