+++ app/src/main/res/values-v31/themes.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <style name="Theme.App.Splash" parent="Theme.SplashScreen">
+        <item name="windowSplashScreenBackground">@color/splash_background</item>
+
+        <item name="windowSplashScreenAnimatedIcon">@drawable/ic_splash_transparent</item>
+
+        <item name="windowSplashScreenAnimationDuration">0</item>
+
+        <item name="android:windowSplashScreenBrandingImage">@null</item>
+
+        <item name="postSplashScreenTheme">@style/Theme.RecipeAndroidProject</item>
+    </style>
+</resources>
