+++ app/src/main/res/values/colors.xml
@@ -1,5 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="black">#FF000000</color>
-    <color name="white">#FFFFFFFF</color>
+    <color name="white">#FFFFFF</color>
+    <color name="black">#000000</color>
+    <color name="primary">#e27036</color>
+    <color name="primary_dark">#E64A19</color>
+    <color name="secondary">#FFC107</color>
+    <color name="splash_background">#FFFFFF</color>
 </resources>
