+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="app_name">recipe_android_project</string>
+    <string name="version_1_0">Version 1.0</string>
+    <string name="discover_delicious_recipes">Discover Delicious Recipes</string>
+    <string name="recipe_app">Recipe App</string>
 </resources>
